Support for additional Appoptics Host Agent Plugins
JIRA: AO-9502

Host Agent in container now has support for:
*  Apache
*  Docker
*  Elasticsearch
*  Kubernetes
*  Mesos
*  Mongodb
*  Mysql
*  Rabbitmq
*  Zookeeper
*  Psutil

Each plugin can be enable by setting 'APPOPTICS_ENABLE_<plugin_name>' to 'true'.
In case of Psutil plugin, the flag is 'APPOPTICS_DISABLE_HOSTAGENT' and to turn the plugin, it should be set to 'false'.By default only Docker plugin is turned on.